+++ polaris/ocean/tests/galewsky_jet/galewsky_jet.cfg
@@ -21,45 +21,3 @@ min_pc_fraction = 0.1
 
 # config options for baroclinic channel testcases
 [galewsky_jet]
-
-# the size of the domain in km in the x and y directions
-lx = 160.0
-ly = 500.0
-
-# time step per resolution (s/km), since dt is proportional to resolution
-dt_per_km = 30
-
-# barotropic time step per resolution (s/km), since btr_dt is proportional to
-# resolution
-btr_dt_per_km = 1.5
-
-# Logical flag that determines if locations of features are defined by distance
-# or fractions. False means fractions.
-use_distances = False
-
-# Viscosity values to test for rpe test case
-viscosities = 1, 5, 10, 20, 200
-
-# Temperature of the surface in the northern half of the domain.
-surface_temperature = 13.1
-
-# Temperature of the bottom in the northern half of the domain.
-bottom_temperature = 10.1
-
-# Difference in the temperature field between the northern and southern halves
-# of the domain.
-temperature_difference = 1.2
-
-# Fraction of domain in Y direction the temperature gradient should be linear
-# over. Used when use_distances = False.
-gradient_width_frac = 0.08
-
-# Width of the temperature gradient around the center sin wave. Default value
-# is relative to a 500km domain in Y. Used when use_distances = True.
-gradient_width_dist = 40e3
-
-# Salinity of the water in the entire domain.
-salinity = 35.0
-
-# Coriolis parameter for entire domain.
-coriolis_parameter = -1.2e-4
