--- polaris/ocean/tests/galewsky_jet/test_balance/test_balance.py
@@ -1,60 +1,0 @@
-from polaris.mesh.spherical import IcosahedralMeshStep
-from polaris.ocean.tests.galewsky_jet.cull_mesh import CullMesh
-from polaris.ocean.tests.galewsky_jet.forward import Forward
-from polaris.ocean.tests.galewsky_jet.initial_state import InitialState
-from polaris.testcase import TestCase
-from polaris.validate import compare_variables
-
-
-class TestBalance(TestCase):
-    """
-    A class to define the Galewsky jet test balance test cases
-
-    Attributes
-    ----------
-    resolution : str
-        The resolution of the test case
-    """
-
-    def __init__(self, test_group, resolution, long):
-        """
-        Create the test case
-
-        Parameters
-        ----------
-        test_group :
-        polaris.ocean.tests.galewsky_jet.GalewskyJet
-            The test group that this test case belongs to
-
-        resolution : str
-            The resolution of the test case
-
-        long : bool
-            Whether to run a long (3-year) simulation to quasi-equilibrium
-        """
-        name = 'performance'
-        self.resolution = resolution
-        self.long = long
-
-        if long:
-            name = 'long'
-
-        subdir = f'{resolution}/{name}'
-        super().__init__(test_group=test_group, name=name,
-                         subdir=subdir)
-
-        self.add_step(IcosahedralMeshStep(
-        #    test_case=self, cell_width=int(resolution[:-2])))
-            test_case=self, cell_width=120))
-        #self.add_step(
-        #    InitialState(test_case=self, resolution=resolution))
-        #self.add_step(
-        #    Forward(test_case=self, resolution=resolution,
-        #            long=long))
-
-    def configure(self):
-        """
-        Set config options for the test case
-        """
-        config = self.config
-        config.add_from_package('polaris.mesh', 'mesh.cfg')
